Make layout fill screen height.
It would be so nice if a flex-grow would just do this without requiring
all its ancestors to be height 100%.. but that seems to only work in
quirks mode.

Also, setting the body's min-height to 100% and #root's height to 100%
does, unintuitively, not stretch #root. Setting body's height to 100%
fixes that but prevents it from stretching along. Opting for flex layout
as the solution here.