Meddle with flex layout and scrolling
Use flex-boxes with column direction and 'flex-grow: 1' all the way from body to
content, so the content always fills the screen, and the footer is also at the
bottom when content is short. Body gets a fixed height of 100vh (would be
min-height if we wished it to stretch&scroll, but see below).

Fixate the width of the InfoPane.

Make ContentPane scroll, instead of the body. We want this mainly for horizontal
scrolling (if needed). However, having vertical scroll done by the body but
horizontal by the ContentPane gives problems:
  1. react-waypoint won't trigger (should be fixable by passing its
     scrollableAncestor prop)
  2. browsers (at least Firefox) don't give a smooth experience: it scrolls
     either one or the other container, and seems to prefer the inner one (which
     would be the horizontal scroll, in our case).
Therefore, let's also do vertical scrolling in the ContentPane for now, with the
downside that the header and footer eat the often precious vertical screen
space. Can hopefully be changed again soon.

Make SearchFilterSchema scroll if too wide. I guess that's what this effort all
started with.. It is definitely not ideal either, but works for now. Blueprint's
CollapsibleList seems less-than-ideal, as it must be told exactly how many items
to show, rather than measuring available space; also it seems to require
MenuItems in the dropdown menu, rather than allowing a free-form content.