Fixed issues with sidebar icons
Changed from using Blueprint SVG JS API to CSS font as they were causing a problem on some platforms/browsers (but oddly not others), such as not rendering at all, rendering with incorrect positioning and minor animation issues (such as with tab interaction animations).

SVG fonts are new in Blueprint 2.0 and although the CSS fonts are to present, but deprecated, the font currently works better so am using them instead for now.

The documentation on how to use them is a little confusing and has seems to make a conflicting suggestion about how they should/shouldn’t be used directly (e.g. via <Icon> or not), possibly a reflection of them being a work in progress.