Special case handling for Excel files
This avoids the old behaviour of rendering two views when there is a spreadsheet with sheets (which was more visible as a bug now that views have toolbars).

This avoids using the Table view for Excel files that have children (i.e. that contain sheets); only the list of each sheet will be displayed (these have the extension csv and open normally).