Improved views (toolbars, paging, width limits…) (#311)
* Adding dotenv for easier app config

This is intended to make local development easier by avoiding having to set the the API endpoint in an env var before running`npm start`.

* Adding tabs to document info sidebar

* Unifying Entity view and Document view

* They use seperate sidebar renderers, but components with in the sidebar are common (e.g. for displaying collection info and related tags).
* The Entity Content view is work in progress. It’s trying do a better job of surfacing the info people would want looking at an entity. It doesn’t look very nice yet but it’s functional.

* Removing debug statement

* Improvements to Document/Entity Sidebar

Some minor improvements to yesterdays work.

* Added icon above name as a cue for the type of thing you are looking at.

* Added “Overview” to Entity sidebar to show number of relationships.

Both of these are in different ways duplication of data, but it is consistant and helpful, especially if the sidebar for either is used in different views (e.g. on a selected element in a search).

Additional:

* Minor update up of how relationship information is displayed on the Entity page so that it’s consistant between the page and the sidebar and other parts of the app (with similar heading and table styles).

* Sidebar gracefully handles no related rags

* Refactoring Toolbar

* Removed Toolbar from Document Content view.
* Added Toolbar to Folder viewer.
* Added Toolbar to PDF viewer.
* Refactored page controls for PDF viewer so they display in Toolbar.

Note: The toolbar is now not displayed in other views. It will be added back to each viewer with affordances for the type of content being displayed.

* Improved PDF viewer

* Fixed issues with sidebar icons

Changed from using Blueprint SVG JS API to CSS font as they were causing a problem on some platforms/browsers (but oddly not others), such as not rendering at all, rendering with incorrect positioning and minor animation issues (such as with tab interaction animations).

SVG fonts are new in Blueprint 2.0 and although the CSS fonts are to present, but deprecated, the font currently works better so am using them instead for now.

The documentation on how to use them is a little confusing and has seems to make a conflicting suggestion about how they should/shouldn’t be used directly (e.g. via <Icon> or not), possibly a reflection of them being a work in progress.

* Improved Document Viewers

* New PDF Viewer
* New Text Viewer
* New HTML Viewer
* New Email Viewer
* New Folder Viewer
* New Table Viewer

All viewers have a toolbar now.

Search is disabled on all viewers for now as the old search mechanism has stopped working.

* Special case handling for Excel files

This avoids the old behaviour of rendering two views when there is a spreadsheet with sheets (which was more visible as a bug now that views have toolbars).

This avoids using the Table view for Excel files that have children (i.e. that contain sheets); only the list of each sheet will be displayed (these have the extension csv and open normally).

* Minor improvement to paging controls

Handles loading more gracefully by not displaying when paging data is not available yet.

* Minor syntax change to remove warning

* Restoring icon/schema for doc info

This was removed and set to empty for some reason in #04bcb110

* Removing cruft

Accidentally committed tmp merge files left behind when running a file merge to resolve conflicts previous commit.

* Improved error handling when PDF fail to laod

* Improved Entity View

Adding header to improve quality of entity view. This is a minor cosmetic change but significantly approves the apperance of entities.

* Improved paging control in Toolbar

Now says “Page X of Y” instead of just “X of Y” and uses react-intl (as a hold over from refactoring, it didn’t previously).

* Refactoring Paging Buttons and CSS

* Paging Buttons moved to dedicated component.
* Refactored to page scoped CSS (and removed old cruft).
* Refactored custom muted class to Blueprint text util.

* Fixing warning about unused const

* CSS Refactoring

* Refactoring Entity CSS

* Refactoring bind based on request in #311

* Refactoring CSS to use Blueprint box shadow

* Refactoring to remove content width helper

Now rolled in to .ContentPageOuter/.ContentPaneInner elements which I think is the only place we need this now it’s been refactored in a common set of classes used by both the Document and Entity views.

* Making footer lighter

Request from @pudo as the darker footer is too promimently displayed when data is loading and is distracting.

* Centering footer text