Moved search bar to top navigation 😱
This appears to work and not be horribly broken.

Most of the changes are related to making the Search components not freak out and just render gracefully if they don’t get passed a SearchContext prop (as we are only doing that on the Search page right now).

As a follow up to this, we plan to refactor to move the PageNavbar and Footer from PageLayout into Screen, and then also move SearchContext out of page PageLayout and into Screen as well.

I wanted to do handle these changes incrementally, first global search, then refactoring of things into Screen, so those changes - which will also take care of cleaning up cruft - will follow.