Minor tweaks to Preview
* Removed box shadow when Preview is hidden so it doesn’t cast shadow on page when it shouldn’t be visible.

* Preview close button now rounded.

This matches the style used for the maxmise toggle and the ‘X’ used when filtering by facets. It’s intended to set a tone for a consistent style for close buttons, and to make it visually distinct from other actions on the tooglebar.

It might be worth making it a class - or fixing pt-round class so that it is actually able to style buttons like other utility classes.