Links to Previews now open conditionally!
If you set `<Collection.Link preview={true}>` or `<Entity.Link preview={true}>` a link which opens in preview sidebar will be generated.

If you don’t set the preview option (or set it to something falsey) then a normal link will be created.

This allows us to link to collections and entities (which includes document in this context) in the Preview on the `Search` and `Collections` screens, by passing a showLinksInPreview prop down - and without impacting how the links behave in other views.

e.g. Clicking a link in the folder view now opens the file or folder being linked to normally again.