Upgrading to react-pdf 3.0.0-beta to fix crashing bug
Upgradeing to react-pdf 3.0.0-beta which resolves the issue where react-pdf crashes when loading in a new PDF in preview mode.

The author did create a fix for the 2.x branch to address the specific bug we are running into, but it didn’t appear to work for us (even with quite abit of futzing) but the upgrade to 3.0 was painless. When 3.0 is released we should upgrade to the final release (this is expected to be soon).

For more info see: https://github.com/wojtekmaj/react-pdf/issues/145