Fix for ‘Show more’ button generating URLs
Because JavaScript is terrible, clicking the ‘Show more’ button was generating URLs like:
`/search#facet:collection_id=1010`

Instead of:
`/search#facet:collection_id=20`

I ran into the same bug handling page numbers with the pagination control so was familar. It uses `parseInt(n, 10)` to tell JavaScript to use base 10 because otherwise parseInt defaults to using one of three different bases depending on what it *thinks* the string is in (based on it’s value).