Navbar heading in bold
This overrides the blueprint framework style for this component.

This override was previously removed as it was particularly distracting in the UI when we had a long name for the site, but I’ve been asked to make it bold again. We how have a shorter name for the production service so it’s not as problematic to have it in bold.