Fix for PDF view sometimes being the wrong size
The onLoadSuccess fires before the document is actually rendered.

Sometimes (but nto often) this results in the first page viewed being rendered too small (it then pops to the right size when you view the next page). This is rare but can happen when a PDF is slow to load. This is a bit of a hack to work around it by waiting 1 second after loading then just double checking if we have calculated the width already.