Improved PDF Viewer resize handling
Fixed bug introduced by using flex layout whereby the view would grow but not shrink on page resize.

The primary fix is to have .out use width: 100% and .inner to hide overflow and move the #PdfViewer element ID back up to the inner element so we can resize the canvas to always fit inside it.

The react-pdf Page canvas can’t take a % value (e.g. 100%), it only works with absolute values, which is why we have to do chicanery with outer elements that are allowed to be fluid, then check their absolute size (in px) and pass that to the canvas layer.