Workaround for document search lifecycle bug (?)
We think there is maybe a lifecycle bug in how we get search results for document search, which might be a wider problem.

After calling queryDocumentRecords for a *new* search we never actually get the results passed in props so the new search result gets stuck in ‘loading’. Searching for something else, then searching *again* for the previous term works - i.e. it did update the query and cache the result, but for some reason it never gets passed as an updated prop.

This somewhat esoteric workaround resolves it by storing the state in a prop, using a fetchRecords boolean in state that we reset whenever we receive new props and using debounce() to make sure updateSearchResults is  not called to often (as it’s trigged by componentDidUpdate()).

I appreciate this is not easy to follow and we need to actually investigate the cause of the underlying issue.