Document View search mostly working but incomplete
Good:

* Document Search in Document View get the default queryText value from the query string.
* It update query string correctly when a new document search is performed (just like it does in Preview).

Bad:

* [BUG] Document Search results in Document View do NOT work correctly yet (but they do in Preview).
* Code / flow is hard to follow (could be simplified, better naming, better inline comments).
* The UI isn’t quite as planned, but is that is secondary to fixing bugs and better left till they are resolved.

The issue wih the bug is that URL updates correctly but the component does not re-render.

We can either try and figure out a workaround (an ugly quick fix for the specific bug, but code and flow will still be hard to follow) or take more time and figure out a better way to do it (and delay the feature release) or decide we want to do both.

Notes:

The following bits of cleanup work should be done before release (once decision about bug made):

* [MINOR] This still includes cruft, especially redundant view code in PdfViewer, that is due to be removed.
* [MINOR] There are a couple of strings that have not been locaized yet (‘Page X’, ‘ No results.’ messages).