Adding Blueprint icons to Sketch file
Add a page that has all the Blueprint SVG icons (using the standard 16px variants), as there is no existing Sketch file for them and this might be the easiest way to use them.

Includes link to origional repo and link to Apache-2.0 licence info to state where they are from.