Added `sample-entities` aleph command
This command is meant to serve any modeling needs by providing random
samples of data from aleph. The data can be filtered by whether it comes
from a secret collection or not and we can subsample and fix the random
seed. There is also filtering on entity schema in addition to which
properties must exist in the entity.

```
$ aleph sample-entities --help
Usage: aleph sample-entities [OPTIONS] [OUTFILE]
  Sample random entities
Options:
  --secret BOOLEAN     Whether to sample from secret collections (None means
                       sample from both)
  -p, --property TEXT  Entities must have at least one of the listed
                       properties
  -s, --schemata TEXT  Filter schematas
  --seed INTEGER       Set the random seed
  --sample-pct FLOAT   Random sampling percent (value from 0-1)
  --limit INTEGER      Number of entities to return
  --help               Show this message and exit.
```