Make sure sorting by undefined field doesn't leave out entities
When we sort by a field that may be undefined in certain schemas, by
default ElasticSearch will  leave out the entities where the field is
undefined from the sorted results.

We use the `unmapped_type` argument to force ElasticSearch not to leave
out those entities from the sorted search results.

Fixes #1857