Add support for Oauth2 Bearer Tokens
**Context**
In order to call Alephs REST endpoints using a JWT token,
support for "Bearer" tokens needs to be added. With it,
Aleph will be able to serve requests from other clients
that use OIDC with the same Identity Server.

**Solution**
The changes are somewhat similar to the logic present in
the OAuth callback endpoint. Aleph tries to determine the
Role based on the foreign ID derived from the JWT token fields
but if it doesn't find one it will create it.
There is no support for groups yet since I haven't found the
need just yet, later tests with an actual Aleph setup and a
configured Indentity Server may uncover that they're needed.