Fix infinite loading issues
We had cases where the API would return inconsistent responses (`result.next` is set and `result.total` > 0, but the actual list of results is always empty for any page) which would result in infinite requests for the subsequent page.

This is an additional safeguard to stop requesting additional pages if the returned list of results is empty.