Fix worker target in Makefile (#2317)
For some reason (we’re not exactly sure why) depending on machine, the aleph CLI is either installed in `/usr/bin/aleph` or in `/aleph/local/bin`. There doesn’t seem to be an easy way to run console_scripts with debugpy.

This was the easiest workaround to start the Aleph CLI with debugpy without specifying an absolute path to the CLI install location.