Always link to entites referenced in entity tables
Previously, we only linked to referenced entities in case the *referencing* entity is a thing.

As discussed in #2338, we actually considered this a bug. The original intent of this condition was probably to check if the *referenced* entity is a thing. Non-thingy entity views do not offer a lot of additional information and are mostly blank, so it doesn’t make a lot of sense to link to them.

That doesn’t seem something that’s possible right now with FtM – referenced entities are always things (see comment https://github.com/alephdata/aleph/issues/2338#issuecomment-1170420470), so I devided to simply render links to referenced entities in all cases.