Increase number of highlight fragments
The default is 3. It’s likely that there will be documents that contain certain search queries more than 3 times on a single page.

Highlight fragments will have a default length of ~120 characters. A standard text page contains about 1500-2000 characters. 15 fragments should cover all occurences of a search query in most cases, even if the frequency of the query on a page is very high.