Fix search results text
Because we do not clear the search when navigating from the search result to a page, the `ResultText` component tried to render a result text for the `searchQuery`. This would result in “Searching…” being displayed left to the search box, even though no search request was pending.

When a single page is displayed, the page data isn’t fetched using a the `searchQuery`, but using the `pageQuery`.