Set the `max_analyzed_offset` setting to prevent ES errors (#2474)
ES errors out when the length of a result body is larger than the
global limit for highlight analysis (defined by
index.highlight.max_analyzed_offset). We set max_analyzed_offset to be
lower than the global setting in highlight queries to prevent those
errors.

The other option was to increase the global limit. But that puts
more pressure on the index.