Remove options to embed diagrams, export as .ftm
Maintaining the option to create network diagram embeds takes a lot of resources and is not a core feature of Aleph. As this feature hasn’t been used a lot, we’re considering to permanently remove it from Aleph. As a first step, we’re only removing the option to create new embeds from the UI, so we can easily revert the change if necessary.

Exporting network diagrams as .ftm files has been necessariy when Aleph Data Desktop was still maintained. Data Desktop is now deprecated, so we’re removing this option as well.

As there is now only a single export option (SVG) left, I have completely removed the export dialog. Instead clicking the export button will now directly download an SVG file.