Simplify and refactor Blueprint overrides
This removes a couple of overrides that have no effect or that are not relevant (e.g. we do not use large or vertical tab styles).

This also replace magic numbers with variables Blueprint uses internally for overrides (e.g. `$menu-item-padding`). This isn’t optimal, as those variables are considered private implementation detail, but it should help keeping things consistent when upgrading Blueprint.

Changes affect the following Blueprint components:

* Tag inputs
* Tags
* Tab lists
* Menu items,
* Buttons