Handle table rows with incorrect number of column
When ingesting CSV files (and probably also other tabular file formats), the resulting CSV file that’s stored in the archive contains a new line file ending.

When we request and parse the CSV file in the frontend, the CSV parser we use emits only a single empty cell for the last row. When we were trying to render other cells in that row, the cell value would always be "undefined" and a skeleton loader would be displayed indefinitely.

Instead, we now compare the number of rows that have been already loaded with the row index of the cell to be rendered.