Refactor: Introduce `TimelineItem`, a wrapper class for entities
`TimelineItem` encapsulates logic commonly required when displaying entities in a timeline, e.g. determining which color to use or what the earliest and latest date the entity covers are. This is in preparation for a new chart view, to ensure common display logic can be shared with the existing list view.