Refactor: Reduce use of optional props
Initially, I kept types a little more relaxed, allowing component consumers to omit certain event handlers. This is handy especially in tests. For example, a test case may not be dependent on passing a real `onChange` handler to a component. When rendering the component in that test, it's convenient to simply omit the `onChange` property rather than passing a no-op event handler.

However, this requires at least some conditional logic in the components. This logic exists only to make writing tests easier. In order to keep the components and prop types as simple as possible, I've marked all props required that are only option for testing purposes. In order to avoid cluttering tests with the same boilerplate props over and over again, I've simply created `defaultProps` objects that can be spread into the components.