Use full entity ID when creating new timeline items
This fixes a bug where creating and removing an entity (without reloading the page between both actions) would result in an error. When a user creates a new entity in a timeline, an ID for the entity is auto-generated. When creating a new entity, Aleph appends a HMAC signature (based on dataset ID and entity ID). The auto-generated ID and the signature form the full ID. In order to remove an entity, the full ID needs to be supplied.

Previously, when creating a new timeline item, only the auto-generated ID was stored. As a consequence, trying to remove an entity using this ID failed. Now, when creating a new item, we wait for the API response that includes the full entity ID and use that for subsequent requests.