Do not save timeline layout on initial load
The current implementation was using `useEffect` to trigger an API request (via the `onLayoutUpdate` callback) whenever the layout state changes. As `useEffect` will execute the side effect on initial render, this caused an undesired API request.

Instead of relying on `useEffect` to persist the layout state, we now trigger the `onLayoutUpdate` callback explicitly when the user performs certain actions.