Add padding/before after earliest/latest date
Previously, if a timeline contains only a single event or all events fell into a very short period, the grid/labels would be rendered for a single month/year only. This looks a bit weird and can also be slightly confusing, so this change now displays a couple of months/years before the earliest/latest date (depending on zoom level).