Move actions to breadcrumbs bar
This saves screen real estate and simplifies the UI (which otherwise would have even more stacked bars at the top of the page). This however means that the timeline actions and the actual timeline are rendered in different component subtrees and we can’t simply pass state down and events up using props/callbacks. Instead, we use a context to make the state and action provider available to both components.