Prevent status badge from shifting other breadcrumb elements around.
When there is a status change (for example from "Saved" to "Saving…) the width of the status badge changes. In the timelines screen, this will slightly change the position of the timelines actions (which are position in the center of the breadcrumbs bar). This change prevents the status badge from changing its width after the initial render.