Fix browser compatibility with translucent background in timeline item list
CSS doesn’t (yet) support altering the alpha channel of HEX colors. That’s why we’ve been using a hack to dynamically apply a translucent background color for timeline list items: instead of actually setting the background color, we rendered a pseudo element with a fully opaque background, then set `opacity` on that pseudo element.

This relied on absolutely positioning the pseudo element within the list item, and the item having a relative position. However, `position: relative` for table rows is undefined per the CSS spec, and it doesn’t render in the intended way in WebKit/Safari.

Instead, we now simply calculate a HEX color in JS land that approximates how the color would be rendered at a certain opacity on a white background.