Fix search highlights disappearing when opening entity previews (#3093)
* Merge new object data with previous data

When searching, we keep the search results in the Redux store in a structure similar to the following:

```json
{
  "entities": {
    "abc123": {
      "id": "abc123",
      "schema": "Company",
      "highlight": "<em>Twitter</em> is an online social media and social networking service."
    }
  },
  "results": {
    "entities?filter:schemata:Company&highlight=true&limit=30&q=twitter": {
      "total": 1,
      "page": 1,
      "results": ["abc123"]
    }
  }
}
```

After clicking on one of the search results, we load detailed entity data and use the detailed data to replace the (partial) entity data from the initial search request. However, the response from the second request does not contain the search highlight, i.e. the highlight is removed from the state when loading additional entity data.

We can work around this by merging data instead of replacing data. It does still feel a bit like a hacky workaround, but a proper solution would require making significant changes to our state management solution (probably replacing it with an off-the-shelf solution).

Fix #2268

* Adjust text highlight color to support light and dark backgrounds

When selecting/clicking on a result on a search results page, we change the background color to dark blue.