Add HintPopover component
We have started adding small one-time popover messages when we introduce UI changes to make users aware of them. (For example, we’ve done that for bookmarks and the new timelines.) When a user dismisses the popover, we persist this decision in local storage.

While the logic is quite easy enough to implement over and over again, this component abstracts it into a separate component (displaying a popover, allowing users to dismiss it permanently).