Version bump
The versions of the following packages were increased to the latest versions: `followthemoney`, `followthemoney-store`, `servicelayer`. Due to the fact that all 3 libraries use a newer version of `SQLAlchemy`, the versions of the following packages were also bumped to latest: `Flask`, `Flask-SQLAlchemy`, `SQLAlchemy`, `Flask-Babel`.

The commit also includes minor refactorings necessary as a result of these version bumps.
* `core.py` was refactored due to the `Flask-Babel` version bump [as per the docs](https://python-babel.github.io/flask-babel/#configuration)
* `migration.py`, `util.py` and `archive_api.py` were refactored due to the `SQLAlchemy` version bump