Limit concurrency of the deploy job only
We actually want individual docs PRs to trigger the build job. And we don’t care if there are multiple build jobs running at the same time. But we want there to be only one deploy job at the same time (in addition to the deploy job only being triggered manually or by pushes to the main branch).