Stringify sanitized HTML as HTML (not XML) (#3575)
The `etree.ElementTree.tostring` method stringifies as valid XML by default. While XML allows self-closing tags by default, only some tags can be self-closing in HTML.

This causes problems when rendering the sanitized HTML. Consider this example:

```html
<noscript><script src="..."></script></noscript>
<h1>Page title</h1>
<p>Some text</p>
```

When converting the sanitized element tree to XML, this would result in something like this:

```html
<noscript />
<h1>Page title</h1>
<p>Some text</p>
```

`noscript` is not a valid void element which results in browsers treating all following elements as children of the `noscript` tag, parsing the HTML into an effective DOM structure like this:

```html
<noscript>
  <h1>Page title</h1>
  <p>Some text</p>
</noscript>
```

As a result the actual page contents are hidden when rendered in a browser (if JavaScript is enabled).