Fix default Dockerfile command
I broke the default Dockerfile command in `7cfcb5`. While the Dockerfile did always specify the WSGI app `aleph.manage:app`, this has been incorrect for some time and it should have always been `aleph.wsgi:app`. Previously we never actually ran a container with the default command though and instead specified the command in the `docker-compose.yml` file. I removed the "duplicate" command in the `docker-compose.yml` in the referenced commit and thus the default command in the Dockerfile is now used.

I’ve now changed the default command specified in `Dockerfile` to be in line with the command previously set in `docker-compose.yml` (https://github.com/alephdata/aleph/blob/3.15.5/docker-compose.yml#L76).

* The WSGI app is now `aleph.wsgi:app` (this is specified in `gunicorn.py` not via a CLI option).
* Uses 6 workers by default (instead of 5).

Closes #3606