Update mentions of `master` branch to `main`
We recently renamed the `master` branch to `main` and changed the default branch in GitHub to `develop`. Because a branch with the name `master` doesn’t exist anymore, GitHub redirects all links to the default branch which is not `develop`.

In some cases this leads to problems, e.g. the installation guide effectively links to the Docker Compose configuration and environment templates on the `develop` branch instead of the stable `main` branch.