Fix default Dockerfile command (#3608)
* Fix default Dockerfile command

I broke the default Dockerfile command in `7cfcb5`. While the Dockerfile did always specify the WSGI app `aleph.manage:app`, this has been incorrect for some time and it should have always been `aleph.wsgi:app`. Previously we never actually ran a container with the default command though and instead specified the command in the `docker-compose.yml` file. I removed the "duplicate" command in the `docker-compose.yml` in the referenced commit and thus the default command in the Dockerfile is now used.

I’ve now changed the default command specified in `Dockerfile` to be in line with the command previously set in `docker-compose.yml` (https://github.com/alephdata/aleph/blob/3.15.5/docker-compose.yml#L76).

* The WSGI app is now `aleph.wsgi:app` (this is specified in `gunicorn.py` not via a CLI option).
* Uses 6 workers by default (instead of 5).

Closes #3606

* Update mentions of `master` branch to `main`

We recently renamed the `master` branch to `main` and changed the default branch in GitHub to `develop`. Because a branch with the name `master` doesn’t exist anymore, GitHub redirects all links to the default branch which is not `develop`.

In some cases this leads to problems, e.g. the installation guide effectively links to the Docker Compose configuration and environment templates on the `develop` branch instead of the stable `main` branch.