Fix documentation links (#3667)
* Update documentation links to use docs.aleph.occrp.org domain

When we updated the docs site, we set up redirects for all of these links, so they were still working, but it is time to properly update the links everywhere.

* Remove "Cross-referencing" docs link on investigations homepage

In the new user guide, there is no separate page for cross-referencing anymore. Instead, cross-referencing is mentioned as part of the docs article "Generating entities from a spreadsheet", so this link is redundant.

* Update user guide link in entity mapping UI

Previously, this linked to the developer docs about mappings. As this link is displayed in the Aleph UI, it makes more sense to link to the end user guide.