Do not allow adding immutable entities to entity sets
While it can be useful to add immutable entities (such as documents) to entity sets, entity sets have never really supported that. The entire UI side of entity sets assumes that entity set items can be mutated by users. This led to inconsistencies and bugs. For example, adding a file to an entity set would remove the `contentHash` property, and once added to an entity set, users were able to change properties of files such as the file name.