WIP Filter expand API response by schema
When expanding an entity in a network diagram, we need to be able to include only non-file entities. This change adds a `schemata` query parameter to the entities expand API endpoint, the endpoint used by the network diagram UI, to enable this.

This change is a little hacky and adds even more complexity to the `expand_proxies` method which already is quite complex. I'm not entirely sure this is a good idea.