Fix end-to-end tests in CI (#3746)
* Ensure that end-to-end test run in CI

The `docker compose run` command doesn’t support the `--build-arg` argument. Depending on the Docker Compose version installed, using the `--build-arg` argument would either exit with an error code or simply print the command help.

To achieve the desired outcome, we can first build the image with the build argument, then start the container. This has the additional advantage of separating the build logs from the logs of the actual command execution in the GitHub Actions UI.

* Pull and start services before running tests

This separates logs related to pulling/starting service containers from the logs of the actual test command.