Import PDF.js worker as asset to enable cache busting
The PDF.js worker is dynamically loaded by PDF.js at runtime. The version of PDF.js in the main app bundle and of the worker have to be in sync. Previously, we simply copied the worker file to a static location (`/static/pdf.worker.min.js`). When updating PDF.js, this might cause problems if the worker file is still cached by browsers.

Instead of manually copying the file, we now make use of Webpack’s support for external assets[1]. Webpack copies the file to the build directory during the build step. It also appends a hash based on the file contents to the file name, e.g. `pdf.worker.min.2b1fcb1d45631a4dd93a.js` to ensure that browsers do not load a cached version when the worker has been updated.

[1]: https://webpack.js.org/guides/asset-modules/#url-assets