Remove bookmarks migration
When we first implemented bookmarks as an experimental feature, they were stored in local storage. People started using the feature so we decided to continue working on bookmarks to build it out into a mature feature.

Most importantly, this included storing bookmarks on the server-side in order to support growing numbers of bookmarks per user and to prevent users from losing their bookmarks when clearing browsing data/switching devices.

We also needed to migrate bookmarks that were currently stored locally to the server side. That means the code base had quite some code that only handled migrating the "legacy" local bookmarks to the server. As it’s been a while since we started migrating to server-side bookmarks, it should be safe to remove that migration logic now.

Fix #2864