Completely remove network diagram embeds feature (#3751)
Almost 1.5 years ago we have disabled the UI option to create new embeds (see 0d69cad). Back then, we decided to hide the UI in the first step, but keep the code in place, just in case we needed to bring back the feature in case more users were relying on it than expected.

We are now ready to entirely remove all code related to this feature. As a follow-up to this, we should also remove the `publish_file` method in servicelayer.