Use absolute date/time instead of relative times
With relative times, when looking at a full status page, there will be a lot of numbers that are constantly updated. For example, even if "last updated" hasn’t changed, how it is rendered will change over time (from "last updated now" to "last updated 5 seconds ago" etc.) In addition, the different lengths of the rendered strings can cause layout shifts/line breaks. This makes it more difficult to track a specific row and to identify values that have actually changed. Instead, I’m now displaying the absolute time (if the day is today) or the absolute date and time. This leads to a slightly less cluttered and noisy UI.