Deploy docs from develop branch (instead of main)
In order to publish documentation changes, they currently need to be merged directly into main (which is only possible if they do not contain non-docs changes). Merging documentation changes into develop means that they will only get publish on the next release.

In most cases, documentation changes are enhancements that are not coupled to a release and should be published as soon as possible, and deploying from develop accounts for this use cases.